# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Saas Push Giant API
  version: 0.1.0
  description: Draft API contract for the PWA and Web Push platform.
servers:
  - url: https://api.push.example.com
paths:
...
  /v1/projects/{projectId}/config:
    get:
      summary: Load public project config
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Public project config
...
  /v1/subscriptions/upsert:
    post:
      summary: Create or update a push subscription
      security:
        - ProjectPublicKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpsert"
      responses:
        "200":
          description: Subscription was upserted
        "400":
          description: Invalid subscription payload
...
  /v1/campaigns:
    post:
      summary: Create campaign draft
      security:
        - AdminSession: []
        - ServerApiKey: []
      responses:
        "201":
          description: Campaign created
  /v1/campaigns/{campaignId}/send-now:
    post:
      summary: Queue campaign for immediate sending
      security:
        - AdminSession: []
        - ServerApiKey: []
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "202":
          description: Campaign queued
...
components:
  securitySchemes:
    ProjectPublicKey:
      type: apiKey
      in: header
      name: X-PushGiant-Project-Key
    ServerApiKey:
      type: http
      scheme: bearer
    AdminSession:
      type: apiKey
      in: cookie
      name: pushgiant_session
    WebhookSignature:
      type: apiKey
      in: header
      name: X-PushGiant-Signature
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CampaignId:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    SubscriptionUpsert:
      type: object
      required:
        - project_id
        - endpoint
        - keys
      properties:
        project_id:
          type: string
          format: uuid
        subscriber_id:
          type: string
          format: uuid
        anonymous_id:
          type: string
        external_customer_id:
          type: string
        endpoint:
          type: string
          format: uri
        keys:
          type: object
          required:
            - p256dh
            - auth
          properties:
            p256dh:
              type: string
            auth:
              type: string
        content_encoding:
          type: string
        platform:
          type: string
        browser:
          type: string
        locale:
          type: string
        timezone:
          type: string